+++ src/backend-libs/fightbulc/moment/src/FormatsInterface.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Moment;
+
+/**
+ * Interface FormatsInterface
+ * @package Moment
+ */
+interface FormatsInterface
+{
+    /**
+     * @param string $format
+     *
+     * @return FormatsInterface
+     */
+    public function format($format);
+
+    /**
+     * @param array $customFormats
+     *
+     * @return FormatsInterface
+     */
+    public function setTokens(array $customFormats);
+}