+++ src/backend-libs/fightbulc/moment/src/MomentException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Moment;
+
+/**
+ * MomentException
+ * @package Moment
+ * @author Tino Ehrich (tino@bigpun.me)
+ */
+class MomentException extends \Exception
+{
+}