+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Breaking Changes
+- added $config['locale'].  See config.sample.php - you have to set it.
+
+### Changed
+- new layout & design with more whitespace and more explanations.  
+
+
 ## [0.2.0] - 2018-06-16
 
 ### Changed